using single formatter for both parallel_cucumber and cucumber
